backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Validation and Serialization
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Image Processing
pillow>=10.0.0
python-barcode>=0.15.1
qrcode[pil]>=7.4.2

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.26.0

# For future S3 migration
boto3>=1.34.0

